Small cleanups
leaving a lot of the commented bad/nonfunctioning or otherwise not great code in this for learning and because there is an unexpected behavior if one comments out 117, 118, 119, 150, 151, 152 and uncomments 153,154,155.  That behavior being that only oscilators persist BUT when 150, 151, 152 are uncommented while 153, 154 and 155 remain commented, the oscilators die out with everything else.