A rather simple "gallery website" powered by Python
It will likely will require a lot of work to get it running on other computers and/or actual servers.  It's made to demonstrate basic understanding of HTML, HTML forms, CSS, serverside scripting (Python here of course), use of SQL and stitching them all together into something usable.  There is JS involved as well but I can't claim any of it as my own work.